Create QR-codes generation screen

+++ core/database/src/main/java/com/arshapshap/versati/core/database/model/qrcodesfeature/QRCodeRequestLocal.kt
@@ -10,8 +10,8 @@ data class QRCodeRequestLocal(
     @ColumnInfo(name = "id") val id: Long,
     @ColumnInfo(name = "data") val data: String,
     @ColumnInfo(name = "size") val size: Int,
-    @ColumnInfo(name = "color") val color: Long,
-    @ColumnInfo(name = "background_color") val backgroundColor: Long,
+    @ColumnInfo(name = "color") val color: Int,
+    @ColumnInfo(name = "background_color") val backgroundColor: Int,
     @ColumnInfo(name = "quiet_zone") val quietZone: Int,
     @ColumnInfo(name = "format") val format: String
 )
