+++ feature/qrcodes/api/src/main/java/com/arshapshap/versati/feature/qrcodes/api/domain/model/QRCodeOptions.kt
@@ -4,8 +4,8 @@ data class QRCodeOptions(
     val id: Long,
     val data: String,
     val size: Int,
-    val color: Long,
-    val backgroundColor: Long,
+    val color: Int,
+    val backgroundColor: Int,
     val quietZone: Int,
     val format: ImageFormat
 )
